Create scoreCheck function, score variables and fix refresh bug
Score variables keep track of player scores.
Every win increments the score variables.
If either score reaches 5, the winner is printed to the console and the game ends.
fix refresh bug where refreshing the page would add to the players score.
Add if statement to gameCheck player victory section where the score will not increment if the player value is null.